+++ arg_parser.py
@@ -22,6 +22,8 @@
         arg_valid_val
         arg_wildcard
         arg_xor_dict
+        _file_create
+        _parse_multi
 
 """
 
@@ -240,7 +242,8 @@ def arg_file_chk(args_array, file_chk_list, file_crt_list=None):
 
             except IOError as (errno, strerror):
 
-                testme()
+                exit_flag = _file_create(name, x, file_crt_list, errno,
+                                         strerror, exit_flag)
                 # If in create list and not present.
                 #if x in file_crt_list and errno == 2:
 
@@ -597,7 +600,8 @@ def arg_xor_dict(args_array, opt_xor_dict):
     return xor_flag
 
 
-def _file_create(name, option, file_crt_list, errno, strerror, exit_flag):
+def _file_create(name, option, file_crt_list, errno, strerror, exit_flag,
+                 **kwargs):
 
     """Function:  _file_create
 
@@ -613,9 +617,13 @@ def _file_create(name, option, file_crt_list, errno, strerror, exit_flag):
         (input) errno -> Current error status from file_crt_list function.
         (input) strerror -> Current error message from file_crt_list function.
         (input) exit_flag -> Current status of file_crt_list function.
+        (input) **kwargs:
+            None
         (output) exit_flag -> True|False - if file creation fails.
 
     """
+    
+    file_crt_list = list(file_crt_list)
 
     if option in file_crt_list and errno == 2:
 
@@ -636,3 +644,59 @@ def _file_create(name, option, file_crt_list, errno, strerror, exit_flag):
         exit_flag = True
 
     return exit_flag
+
+
+def _parse_multi(argv, args_array, opt_def_dict, **kwargs):
+
+    """Function:  _parse_multi
+
+    Description:  Processes a multi-value argument in command line
+        arguments.  Modifys the args_array by adding a dictionary key and a
+        list of values.
+
+    NOTE:  Used by the arg_parse2() to reduce the complexity rating.
+
+    Arguments:
+        (input) argv -> Arguments from the command line.
+        (input) args_array -> Array of command line options and values.
+        (input) opt_def_dict -> Dict with options and default values.
+        (input) **kwargs:
+            None.
+        (output) argv -> Arguments from the command line.
+        (output) args_array -> Array of command line options and values.
+
+    """
+    
+    argv = list(argv)
+    args_array = dict(args_array)
+    opt_def_dict = dict(opt_def_dict)
+    
+    # If no value in argv for option and it's not an integer.
+    if len(argv) < 2 \
+       or (argv[1][0] == "-" and not gen_libs.chk_int(argv[1])):
+
+        # See if default value is available for argument.
+        args_array = arg_default(argv[0], args_array, opt_def_dict)
+
+    else:
+        # Handle multiple values for argument.
+        args_array[argv[0]] = []
+        x = 0
+        tmp_argv = argv[1:]
+
+        # Process values until next argument.
+        while tmp_argv:
+
+            if tmp_argv[0][0] == "-":
+                break
+
+            else:
+                args_array[argv[0]].append(tmp_argv[0])
+
+            x = x + 1
+            tmp_argv = tmp_argv[1:]
+
+        # Move to argument after the multiple values.
+        argv = argv[x:]
+
+    return argv, args_array